Fixes + depoyTips (#199)
* toml2csv: Fix filtering when transformation is active

* toml2csv: manage inputfile encoding + more logs

* More doc for scratch:pool:create

* Toml2CSV updates

- Refresh spinner every 10 secs
- More details in spinner message
- Enhance error messages summary
- Manage required attributes
- Manage truncate values

* deployTips + toml2csv

* Add listViewMine in cleaning references

* Clean MDAPI errors output logs

* Do not add line in cache if failed

* Do not create log files in the current directory if it is empty

* changelog

* [Mega-Linter] Apply linters fixes

* Remove MegaLinter Github status reporter

+++ src/commands/hardis/misc/toml2csv.ts
@@ -82,6 +82,7 @@ export default class Toml2Csv extends SfdxCommand {
   protected doFilterSections = false;
 
   protected spinner: any;
+  protected spinnerInterval: any;
   protected inputFileSeparator: string;
   protected outputFileSeparator: string;
   protected tomlSectionsFileWriters: any = {};
@@ -90,8 +91,11 @@ export default class Toml2Csv extends SfdxCommand {
 
   protected csvFiles: string[] = [];
 
+  protected currentSection: string | null = null;
   protected sectionLineIds: any = {};
   protected sectionLines: any = {};
+
+  protected lineErrorMessages: any = {};
 
   protected stats = {
     sectionLinesNb: 0,
@@ -121,12 +125,12 @@ export default class Toml2Csv extends SfdxCommand {
 
     // Check TOML file is existing
     if (!fs.existsSync(tomlFile)) {
-      this.triggerError(`TOML file ${tomlFile} not found`);
+      this.triggerError(c.red(`TOML file ${c.bold(tomlFile)} not found`));
     }
 
     // Read configuration file
     if (!fs.existsSync(this.transfoConfigFile)) {
-      this.triggerError(`Mapping/Transco config ${this.transfoConfigFile} not found`);
+      this.triggerError(c.red(`Mapping/Transco config ${c.bold(this.transfoConfigFile)} not found`));
     }
     const transfoConfigInit = JSON.parse(fs.readFileSync(this.transfoConfigFile));
     this.transfoConfig = this.completeTransfoConfig(transfoConfigInit);
@@ -138,21 +142,25 @@ export default class Toml2Csv extends SfdxCommand {
     // Create output directory if not existing yet
     await fs.ensureDir(this.outputDir);
     // Empty output dir
-    await fs.emptyDir(this.outputDir);
-    await fs.ensureDir(path.join(this.outputDir, "errors"));
+    if (!this.transfoConfig?.skipResetOutputDir === true) {
+      await fs.emptyDir(this.outputDir);
+      await fs.ensureDir(path.join(this.outputDir, "errors"));
+    }
 
     uxLog(this, c.cyan(`Generating CSV files from ${c.green(tomlFile)} (encoding ${tomlFileEncoding}) into folder ${c.green(this.outputDir)}`));
 
     // Start spinner
     this.spinner = ora({ text: `Processing...`, spinner: "moon" }).start();
+    this.spinnerInterval = setInterval(() => {
+      this.updateSpinner();
+    }, 10000);
 
     // Read TOML file and process lines section by section
-    const fileStream = fs.createReadStream(tomlFile, { encoding: "utf8" });
+    const fileStream = fs.createReadStream(tomlFile, { encoding: this.transfoConfig?.inputFile?.encoding || "utf8" });
     const rl = readline.createInterface({
       input: fileStream,
       crlfDelay: Infinity,
     });
-    let currentSection = null;
     for await (const line of rl) {
       this.stats.totalLinesNb++;
       if (debugMode) {
@@ -166,45 +174,43 @@ export default class Toml2Csv extends SfdxCommand {
       // Section line
       if (line.startsWith("[")) {
         this.stats.sectionLinesNb++;
-        currentSection = /\[(.*)\]/gm.exec(line)[1]; // ex: get COMPTES from [COMPTES]
-        if (this.doFilterSections && !this.filterSections.includes(currentSection)) {
+        this.currentSection = /\[(.*)\]/gm.exec(line)[1]; // ex: get COMPTES from [COMPTES]
+        if (this.doFilterSections && !this.filterSections.includes(this.currentSection)) {
           continue;
         }
-        this.spinner.text =
-          `Processing section ${currentSection} (data lines: ${this.stats.dataLinesNb},` +
-          ` errors: ${this.stats.dataErrorLinesNb}, filtered: ${this.stats.dataFilteredLinesNb})`;
+        this.updateSpinner();
         // Init section variables
-        this.stats.sections[currentSection] = this.stats.sections[currentSection] || {
+        this.stats.sections[this.currentSection] = this.stats.sections[this.currentSection] || {
           dataLinesNb: 0,
           dataSuccessLinesNb: 0,
           dataErrorLinesNb: 0,
           dataFilteredLinesNb: 0,
           dataFilterErrorsNb: 0,
-          dataDuplicatesNb: this.transfoConfig?.entities[currentSection]?.removeDuplicates ? 0 : null,
+          dataDuplicatesNb: this.transfoConfig?.entities[this.currentSection]?.removeDuplicates ? 0 : null,
         };
-        this.sectionLineIds[currentSection] = [];
-        this.sectionLines[currentSection] = this.sectionLines[currentSection] || [];
+        this.sectionLineIds[this.currentSection] = [];
+        this.sectionLines[this.currentSection] = this.sectionLines[this.currentSection] || [];
         // Init section files writeStreams
-        if (this.tomlSectionsFileWriters[currentSection] == null) {
-          this.tomlSectionsFileWriters[currentSection] = await this.createSectionWriteStream(currentSection, false);
+        if (this.tomlSectionsFileWriters[this.currentSection] == null) {
+          this.tomlSectionsFileWriters[this.currentSection] = await this.createSectionWriteStream(this.currentSection, false);
           if (!this.skipTransfo) {
-            this.tomlSectionsErrorsFileWriters[currentSection] = await this.createSectionWriteStream(currentSection, true);
+            this.tomlSectionsErrorsFileWriters[this.currentSection] = await this.createSectionWriteStream(this.currentSection, true);
           }
         }
       }
       // CSV line
-      else if (currentSection) {
-        if (this.doFilterSections && !this.filterSections.includes(currentSection)) {
+      else if (this.currentSection) {
+        if (this.doFilterSections && !this.filterSections.includes(this.currentSection)) {
           continue;
         }
         this.stats.dataLinesNb++;
-        this.stats.sections[currentSection].dataLinesNb++;
+        this.stats.sections[this.currentSection].dataLinesNb++;
         const lineSplit = line.split(this.inputFileSeparator);
         // Check if line has to be filtered
         let filtered = false;
-        if (this.transfoConfig?.entities[currentSection]?.filters) {
-          for (const filter of this.transfoConfig?.entities[currentSection]?.filters) {
-            if (!this.checkFilter(filter, lineSplit, currentSection)) {
+        if (this.transfoConfig?.entities[this.currentSection]?.filters) {
+          for (const filter of this.transfoConfig?.entities[this.currentSection]?.filters) {
+            if (!this.checkFilter(filter, lineSplit, this.currentSection)) {
               filtered = true;
               break;
             }
@@ -212,29 +218,28 @@ export default class Toml2Csv extends SfdxCommand {
         }
         if (filtered) {
           this.stats.dataFilteredLinesNb++;
-          this.stats.sections[currentSection].dataFilteredLinesNb++;
+          this.stats.sections[this.currentSection].dataFilteredLinesNb++;
+          continue;
         }
         if (this.skipTransfo) {
           // Without transformation
-          if (!filtered) {
-            const lineSf = lineSplit
-              .map((val) => (this.inputFileSeparator !== this.outputFileSeparator ? this.formatCsvCell(val) : val)) // Add quotes if value contains a separator
-              .join(this.outputFileSeparator);
-            if (this.checkNotDuplicate(currentSection, lineSf)) {
-              await this.writeLine(lineSf, this.tomlSectionsFileWriters[currentSection]);
-              this.addLineInCache(currentSection, lineSplit, lineSf);
-              this.stats.sections[currentSection].dataSuccessLinesNb++;
-              this.stats.dataSuccessLinesNb++;
-            }
+          const lineSf = lineSplit
+            .map((val) => (this.inputFileSeparator !== this.outputFileSeparator ? this.formatCsvCell(val) : val)) // Add quotes if value contains a separator
+            .join(this.outputFileSeparator);
+          if (this.checkNotDuplicate(this.currentSection, lineSf)) {
+            await this.writeLine(lineSf, this.tomlSectionsFileWriters[this.currentSection]);
+            this.addLineInCache(this.currentSection, lineSplit, lineSf);
+            this.stats.sections[this.currentSection].dataSuccessLinesNb++;
+            this.stats.dataSuccessLinesNb++;
           }
         } else {
           // With transformation
           try {
-            await this.convertLineToSfThenWrite(currentSection, lineSplit);
+            await this.convertLineToSfThenWrite(this.currentSection, lineSplit);
           } catch (e) {
             // Manage error
             this.stats.dataErrorLinesNb++;
-            this.stats.sections[currentSection].dataErrorLinesNb++;
+            this.stats.sections[this.currentSection].dataErrorLinesNb++;
             const lineError =
               line
                 .split(this.inputFileSeparator)
@@ -242,11 +247,16 @@ export default class Toml2Csv extends SfdxCommand {
                 .join(this.outputFileSeparator) +
               this.outputFileSeparator +
               `"${e.message.replace(/"/g, "'")}"`;
-            if (this.checkNotDuplicate(currentSection, lineError)) {
-              await this.writeLine(lineError, this.tomlSectionsErrorsFileWriters[currentSection]);
-              this.addLineInCache(currentSection, lineSplit, lineError);
+            if (this.checkNotDuplicate(this.currentSection, lineError)) {
+              await this.writeLine(lineError, this.tomlSectionsErrorsFileWriters[this.currentSection]);
+              this.addLineInCache(this.currentSection, lineSplit, lineError, false);
             }
-            uxLog(this, c.red(e.message));
+            if (this.lineErrorMessages[e.message]) {
+              this.lineErrorMessages[e.message]++;
+            } else {
+              this.lineErrorMessages[e.message] = 1;
+              uxLog(this, c.red(e.message));
+            }
           }
         }
       } else {
@@ -267,6 +277,7 @@ export default class Toml2Csv extends SfdxCommand {
     }
 
     // Stop spinner
+    clearInterval(this.spinnerInterval);
     this.spinner.succeed(`File processing complete of ${this.stats.dataLinesNb} data lines (${this.stats.dataErrorLinesNb} in error)`);
 
     // Manage file copy to data workspace folders
@@ -283,8 +294,19 @@ export default class Toml2Csv extends SfdxCommand {
       }
     }
 
-    // Display summary results
+    // Display full stats
     uxLog(this, c.grey("Stats: \n" + JSON.stringify(this.stats, null, 2)));
+
+    // Display errors summary
+    if (Object.keys(this.lineErrorMessages)) {
+      uxLog(this, c.yellow("There have been parsing errors:"));
+      for (const errMsg of Object.keys(this.lineErrorMessages)) {
+        uxLog(this, c.yellow("- " + this.lineErrorMessages[errMsg] + " lines: " + errMsg));
+      }
+      uxLog(this, "");
+    }
+
+    // Display human-readable stats
     for (const section of Object.keys(this.stats.sections)) {
       const sectionStats = this.stats.sections[section];
       if (sectionStats.dataLinesNb > 0) {
@@ -298,6 +320,13 @@ export default class Toml2Csv extends SfdxCommand {
       c.cyan(`TOML file ${c.green(tomlFile)} has been split into ${c.green(this.csvFiles.length)} CSV files in directory ${c.green(this.outputDir)}`)
     );
     return { outputString: message, csvfiles: this.csvFiles, stats: this.stats };
+  }
+
+  updateSpinner() {
+    this.spinner.text =
+      `Processing section ${this.currentSection} (total lines: ${this.stats.dataLinesNb},` +
+      ` success: ${this.stats.dataSuccessLinesNb},` +
+      ` errors: ${this.stats.dataErrorLinesNb}, filtered: ${this.stats.dataFilteredLinesNb})`;
   }
 
   // Create output write stream for section
@@ -321,7 +350,7 @@ export default class Toml2Csv extends SfdxCommand {
       // Init writeStream
       const fileWriteStream = fs.createWriteStream(path.resolve(outputFile), { encoding: "utf8" });
       // Create CSV Header
-      let headerLine = this.transfoConfig?.entities[section]?.outputFile?.cols
+      let headerLine = (this.transfoConfig?.entities[section]?.outputFile?.cols || [])
         .map((colDescription: any) => colDescription.name)
         .join(this.outputFileSeparator);
       if (errMode) {
@@ -365,7 +394,7 @@ export default class Toml2Csv extends SfdxCommand {
     const inputCols: any = {};
     if (this.transfoConfig.entities[section]?.inputFile?.cols) {
       // Case when cols are defined line [ {"Name": 0, "FirstName: 1" ...}]
-      for (let i = 0; i < this.transfoConfig.entities[section]?.inputFile?.cols.length; i++) {
+      for (let i = 0; i < this.transfoConfig.entities[section].inputFile.cols.length; i++) {
         const inputColKey = this.transfoConfig.entities[section].inputFile.cols[i];
         inputCols[inputColKey] = lineSplit[i] || "";
       }
@@ -377,15 +406,31 @@ export default class Toml2Csv extends SfdxCommand {
       }
     }
     // convert into output format
-    for (const colDefinition of this.transfoConfig.entities[section].outputFile.cols) {
+    for (const colDefinition of this.transfoConfig.entities[section]?.outputFile?.cols || []) {
       // Col definition is the position or the name of a column in input file
       if (colDefinition.inputColKey || colDefinition.inputColKey === 0) {
         if (inputCols[colDefinition.inputColKey] || inputCols[colDefinition.inputColKey] === "" || inputCols[colDefinition.inputColKey] === 0) {
-          let colVal = inputCols[colDefinition.inputColKey] || "";
+          let colVal: string = inputCols[colDefinition.inputColKey] || "";
           // Transform if necessary
-          if (colVal && colDefinition.transfo) {
+          if (colDefinition.transfo) {
             colVal = this.manageTransformation(colDefinition.transfo, colVal, colDefinition);
           }
+          // Manage missing required value
+          if (colDefinition?.required === true && colVal === "") {
+            this.triggerError(
+              c.red(
+                `${c.bold(this.transfoConfig.entities[this.currentSection].outputFile.salesforceObjectApiName)}.${c.bold(
+                  colDefinition.name
+                )}: Missing required value`
+              ),
+              false
+            );
+          }
+          // Manage truncate value
+          if (colDefinition?.truncate && colVal.length > colDefinition.truncate) {
+            colVal = colVal.substring(0, colDefinition.truncate - 1);
+          }
+          // Add cell in line
           linesSfArray.push(colVal); // Add quotes if value contains output file separator
         } else {
           this.triggerError(c.red(`You must have a correspondance in input cols for output col ${colDefinition}`), false);
@@ -411,7 +456,7 @@ export default class Toml2Csv extends SfdxCommand {
             const colNameValue = linesSfArray[colNamePosition];
             return colNameValue;
           })
-          .join(" ");
+          .join(colDefinition.separator || " ");
         linesSfArray.push(concatenatedValue);
       }
     }
@@ -419,9 +464,13 @@ export default class Toml2Csv extends SfdxCommand {
     // Join line as CSV, as expected by SF Bulk API
     const lineSf = linesSfArray.map((val) => this.formatCsvCell(val)).join(this.outputFileSeparator);
     // Write line with fileWriter
-    await this.writeLine(lineSf, this.tomlSectionsFileWriters[section]);
-    this.stats.sections[section].dataSuccessLinesNb++;
-    this.stats.dataSuccessLinesNb++;
+
+    if (this.checkNotDuplicate(section, lineSf)) {
+      await this.writeLine(lineSf, this.tomlSectionsFileWriters[section]);
+      this.stats.sections[section].dataSuccessLinesNb++;
+      this.stats.dataSuccessLinesNb++;
+      this.addLineInCache(section, lineSplit, lineSf, true);
+    }
   }
 
   // Apply transformations defined in transfoconfig file
@@ -453,7 +502,11 @@ export default class Toml2Csv extends SfdxCommand {
     const transcodedValue = enumValues[colVal] || transfo.default || "";
     if (transcodedValue === "" && colVal !== "") {
       this.triggerError(
-        `There should be a matching value for ${colVal} in ${JSON.stringify(enumValues)} for column ${JSON.stringify(colDefinition)}`,
+        c.red(
+          `${c.bold(this.transfoConfig.entities[this.currentSection].outputFile.salesforceObjectApiName)}.${c.bold(
+            colDefinition.name
+          )}: Missing matching value for ${c.bold(colVal)} in ${c.grey(JSON.stringify(Object.keys(enumValues)))}`
+        ),
         false
       );
     }
@@ -471,12 +524,12 @@ export default class Toml2Csv extends SfdxCommand {
       // Load enum in memory
       const transcoFile = path.join(this.rootConfigDirectory, "enums", `${transfo.enum}.json`);
       if (!fs.existsSync(transcoFile)) {
-        this.triggerError(`Missing transco file ${transcoFile} for enum ${transfo.enum}`, false);
+        this.triggerError(`Missing transco file ${c.bold(transcoFile)} for enum ${c.bold(transfo.enum)}`, false);
       }
       this.loadedTranscos[transfo.enum] = JSON.parse(fs.readFileSync(transcoFile));
       return this.loadedTranscos[transfo.enum];
     }
-    this.triggerError(`Missing transco definition in ${JSON.stringify(transfo)}`, false);
+    this.triggerError(`Missing transco definition in ${c.bold(JSON.stringify(transfo))}`, false);
   }
 
   checkFilter(filter, lineSplit, currentSection) {
@@ -524,8 +577,8 @@ export default class Toml2Csv extends SfdxCommand {
     return res;
   }
 
-  addLineInCache(currentSection, lineSplit, lineWrite) {
-    if (this.transfoConfig?.entities[currentSection]?.idColNumber) {
+  addLineInCache(currentSection, lineSplit, lineWrite, success = true) {
+    if (success && this.transfoConfig?.entities[currentSection]?.idColNumber) {
       const lineId = lineSplit[this.transfoConfig.entities[currentSection].idColNumber - 1];
       this.sectionLineIds[currentSection].push(lineId);
     }
@@ -559,6 +612,7 @@ export default class Toml2Csv extends SfdxCommand {
 
   triggerError(errorMsg: string, fatal = true) {
     if (fatal && this.spinner) {
+      clearInterval(this.spinnerInterval);
       this.spinner.fail(errorMsg);
     }
     throw new SfdxError(errorMsg);
